Update: Modularise TileGrid component
