+++ src/components/core.js
@@ -56,7 +56,7 @@ import Statistic from 'components/alexandria/Statistic/component';
 import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
 import SvgSymbolCircle from 'components/alexandria/SvgSymbolCircleComponent';
 import Tag from 'components/alexandria/Tag/component';
-import TileGrid from 'components/alexandria/TileGridComponent';
+import TileGrid from 'components/alexandria/TileGrid/component';
 import Totals from 'components/alexandria/TotalsComponent';
 
 import 'styles/_bootstrap-custom.scss';
